Extract Windows1255 to its own file
