+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/Windows1255.kt
@@ -1,0 +1,152 @@
+/*
+ *    Copyright 2024 Ondřej Karmazín
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package cz.multiplatform.escpos4k.core.encoding.charset
+
+/** Page 49 [Windows-1255] */
+public data object Windows1255 : Charset(49, "windows-1255") {
+  override val mapping: Map<Char, Int> =
+      mapOf(
+          '\u20AC' to 0x80,
+          '\uFFFD' to 0x81,
+          '\u201A' to 0x82,
+          '\u0192' to 0x83,
+          '\u201E' to 0x84,
+          '\u2026' to 0x85,
+          '\u2020' to 0x86,
+          '\u2021' to 0x87,
+          '\u02C6' to 0x88,
+          '\u2030' to 0x89,
+          '\uFFFD' to 0x8A,
+          '\u2039' to 0x8B,
+          '\uFFFD' to 0x8C,
+          '\uFFFD' to 0x8D,
+          '\uFFFD' to 0x8E,
+          '\uFFFD' to 0x8F,
+          '\uFFFD' to 0x90,
+          '\u2018' to 0x91,
+          '\u2019' to 0x92,
+          '\u201C' to 0x93,
+          '\u201D' to 0x94,
+          '\u2022' to 0x95,
+          '\u2013' to 0x96,
+          '\u2014' to 0x97,
+          '\u02DC' to 0x98,
+          '\u2122' to 0x99,
+          '\uFFFD' to 0x9A,
+          '\u203A' to 0x9B,
+          '\uFFFD' to 0x9C,
+          '\uFFFD' to 0x9D,
+          '\uFFFD' to 0x9E,
+          '\uFFFD' to 0x9F,
+          '\u00A0' to 0xA0,
+          '\u00A1' to 0xA1,
+          '\u00A2' to 0xA2,
+          '\u00A3' to 0xA3,
+          '\u20AA' to 0xA4,
+          '\u00A5' to 0xA5,
+          '\u00A6' to 0xA6,
+          '\u00A7' to 0xA7,
+          '\u00A8' to 0xA8,
+          '\u00A9' to 0xA9,
+          '\u00D7' to 0xAA,
+          '\u00AB' to 0xAB,
+          '\u00AC' to 0xAC,
+          '\u00AD' to 0xAD,
+          '\u00AE' to 0xAE,
+          '\u00AF' to 0xAF,
+          '\u00B0' to 0xB0,
+          '\u00B1' to 0xB1,
+          '\u00B2' to 0xB2,
+          '\u00B3' to 0xB3,
+          '\u00B4' to 0xB4,
+          '\u00B5' to 0xB5,
+          '\u00B6' to 0xB6,
+          '\u00B7' to 0xB7,
+          '\u00B8' to 0xB8,
+          '\u00B9' to 0xB9,
+          '\u00F7' to 0xBA,
+          '\u00BB' to 0xBB,
+          '\u00BC' to 0xBC,
+          '\u00BD' to 0xBD,
+          '\u00BE' to 0xBE,
+          '\u00BF' to 0xBF,
+          '\u05B0' to 0xC0,
+          '\u05B1' to 0xC1,
+          '\u05B2' to 0xC2,
+          '\u05B3' to 0xC3,
+          '\u05B4' to 0xC4,
+          '\u05B5' to 0xC5,
+          '\u05B6' to 0xC6,
+          '\u05B7' to 0xC7,
+          '\u05B8' to 0xC8,
+          '\u05B9' to 0xC9,
+          '\uFFFD' to 0xCA,
+          '\u05BB' to 0xCB,
+          '\u05BC' to 0xCC,
+          '\u05BD' to 0xCD,
+          '\u05BE' to 0xCE,
+          '\u05BF' to 0xCF,
+          '\u05C0' to 0xD0,
+          '\u05C1' to 0xD1,
+          '\u05C2' to 0xD2,
+          '\u05C3' to 0xD3,
+          '\u05F0' to 0xD4,
+          '\u05F1' to 0xD5,
+          '\u05F2' to 0xD6,
+          '\u05F3' to 0xD7,
+          '\u05F4' to 0xD8,
+          '\uFFFD' to 0xD9,
+          '\uFFFD' to 0xDA,
+          '\uFFFD' to 0xDB,
+          '\uFFFD' to 0xDC,
+          '\uFFFD' to 0xDD,
+          '\uFFFD' to 0xDE,
+          '\uFFFD' to 0xDF,
+          '\u05D0' to 0xE0,
+          '\u05D1' to 0xE1,
+          '\u05D2' to 0xE2,
+          '\u05D3' to 0xE3,
+          '\u05D4' to 0xE4,
+          '\u05D5' to 0xE5,
+          '\u05D6' to 0xE6,
+          '\u05D7' to 0xE7,
+          '\u05D8' to 0xE8,
+          '\u05D9' to 0xE9,
+          '\u05DA' to 0xEA,
+          '\u05DB' to 0xEB,
+          '\u05DC' to 0xEC,
+          '\u05DD' to 0xED,
+          '\u05DE' to 0xEE,
+          '\u05DF' to 0xEF,
+          '\u05E0' to 0xF0,
+          '\u05E1' to 0xF1,
+          '\u05E2' to 0xF2,
+          '\u05E3' to 0xF3,
+          '\u05E4' to 0xF4,
+          '\u05E5' to 0xF5,
+          '\u05E6' to 0xF6,
+          '\u05E7' to 0xF7,
+          '\u05E8' to 0xF8,
+          '\u05E9' to 0xF9,
+          '\u05EA' to 0xFA,
+          '\uFFFD' to 0xFB,
+          '\uFFFD' to 0xFC,
+          '\u200E' to 0xFD,
+          '\u200F' to 0xFE,
+          '\uFFFD' to 0xFF,
+      )
+}
